backend/requirements.txt
Django==3.2.7
django-colorfield==0.4.3
django-filter==2.4.0 
djangorestframework==3.11.0 
djoser==2.1.0
drf-extra-fields==3.1.1
gunicorn==20.0.4
psycopg2-binary==2.9.1
reportlab==3.6.2
sqlparse==0.3.1 

